Remove daemon mode parameter
