+++ mock_example.yml
@@ -1,4 +1,3 @@
-daemonMode: false
 consul:
   enabled: false
   consulHost: 127.0.0.1
