+++ src/main/kotlin/ua/com/lavi/komock/KomockRunner.kt
@@ -1,6 +1,5 @@
 package ua.com.lavi.komock
 
-import org.slf4j.LoggerFactory
 import ua.com.lavi.komock.config.ApplicationConfiguration
 import ua.com.lavi.komock.registrar.ConsulRegistrar
 import ua.com.lavi.komock.registrar.ServerRegistrar
@@ -11,8 +10,6 @@ import ua.com.lavi.komock.registrar.SpringConfigRegistrar
  */
 
 class KomockRunner {
-
-    private val log = LoggerFactory.getLogger(this.javaClass)
 
     fun run(applicationConfiguration: ApplicationConfiguration) {
 
@@ -32,18 +29,6 @@ class KomockRunner {
         val consulServerProperties = applicationConfiguration.consul
         if (consulServerProperties.enabled) {
             consulRegistrar.register(consulServerProperties)
-        }
-
-        if (applicationConfiguration.daemonMode) {
-            daemonMode()
-        }
-    }
-
-    private fun daemonMode() {
-        try {
-            Thread.currentThread().join()
-        } catch (e: InterruptedException) {
-            log.warn("Error: {}", e)
         }
     }
 }
