+++ src/main/kotlin/ua/com/lavi/komock/registrar/ConsulRegistrar.kt
@@ -29,5 +29,16 @@ class ConsulRegistrar {
             clientRegistrar.agentServiceRegister(newService)
             log.info("Registered consul service: ${consulService.serviceId} - ${consulService.serviceAddress}:${consulService.servicePort}")
         }
+        daemonMode()
     }
+
+    private fun daemonMode() {
+        try {
+            log.info("Consul registration is running in daemon mode")
+            Thread.currentThread().join()
+        } catch (e: InterruptedException) {
+            log.warn("Error: {}", e)
+        }
+    }
+
 }
